+++ packages/i18n/i18n/it.js
@@ -84,6 +84,7 @@ i18n.map('it', {
 	label_password: "Password",
 	label_privacy: "Privacy",
 	label_location: "Posizione",
+	label_geometry: "Geometria",
 	label_mylocation: "Mia Posizione",
 	label_userblocked: "Utente bloccato",
 	label_histuser: "Luoghi recenti",