+++ packages/import/modules/Util_sanitize.js
@@ -8,13 +8,14 @@ _.extend(K.Util.sanitize, {
 
 		name = name.toLowerCase();
 
-		name = _.str.camelize(name);
+		name = _.str.camelize(name)
+		name = _.str.underscored(name);
 
 		name = name
 			.replace(/[^a-zA-Z0-9\\\-\._]/g,'')
 			.replace(/\.\./g,'.')
 			.replace(/\//g,'')
-			.replace(/[\-\._]/g,'');
+			.replace(/[\-\.]/g,'');
 
 		return name;
 	}